Create CurrencyConversion.py
This is a simple currency conversion application built with Python's tkinter library and the currency_converter module. The user can select the currencies they want to convert from and to using drop-down menus, enter the amount they wish to convert, and click the convert button. The converted currency amount is displayed below. The user can also exit the application by clicking the cancel button.